Updated Node.java, Packet.java and fixed errors

+++ src/com/floodmonitoring/Constants.java
@@ -2,7 +2,7 @@ package com.floodmonitoring;
 
 public class Constants {
 	
-	public static double randomSeed = 0.0;
+	public static int randomSeed = 0;
 	public static double RegularBatteryDecrease=0.01;
 	public static double RecieveBatteryDecrease=0.1;
 	public static double SendBatteryDecrease=0.3;
